requirements.txt
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
docker==7.1.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.14.0
identify==2.5.36
idna==3.7
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
nodeenv==1.9.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydc-control==2.93
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-socks==2.4.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
wcwidth==0.2.13
websockets==12.0
websockets_proxy==0.1.2
zipp==3.19.0
